expose setupCommandTest
