+++ package.json
@@ -68,7 +68,8 @@
     "@types/fbemitter": "^2.0.32",
     "fbemitter": "^2.1.1",
     "flux-standard-action": "^2.0.0",
-    "lodash.merge": "^4.6.0"
+    "lodash.merge": "^4.6.0",
+    "unpartial": "^0.4.1"
   },
   "devDependencies": {
     "@types/node": "^8.0.53",
